requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-heroku==0.3.1
django-import-export==4.3.1
djangorestframework==3.15.2
et_xmlfile==2.0.0
gunicorn==23.0.0
mysqlclient==2.2.6
openpyxl==3.1.5
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlparse==0.5.2
tablib==3.7.0
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
